Added validation and exception handling (3)
In this commit I will include the current progress on data validation and exception handling.
I have started working on the following razor pages:
- Login (done)
- Index (done)
- Register (done)
- Profile (WIP)

I added improvements to the code, data validation wherever it's applicable and exception handling as well, wherever it's applicable.
I also started to work on a way to store the images for the profile picture, as before they were not stored correctly so they wouldn't show up on the Luna site.
I will continue implementation and will do further commits to save my work every time I stop for the day.