Added validation and exception handling (4)
In this commit I will include the current progress on data validation and exception handling.
I have finished working on all the Razor Pages that I have at the moment, changing some functionalities and adding validations in place.
There are still some small issues in the front-end, but nothing that makes the project unusable or broken.

I will continue with revising the classes, removing redundant code and updating the UML accordingly.